feat: add banner to auth pages
