+++ apps/web/src/components/footer/demo-banner.tsx
@@ -1,47 +1,16 @@
-"use client";
-
-import { useTransition } from "react";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { signOutAction } from "@/actions/auth";
-import { toast } from "sonner";
+import { getSession } from "@/lib/auth/auth.server";
 
-import { Button } from "@energyleaf/ui";
+import DemoBannerExit from "./demo-banner-exit";
 
-export function DemoBanner() {
-    const [pending, startTransition] = useTransition();
-    const router = useRouter();
-
-    function signOut() {
-        startTransition(() => {
-            toast.promise(signOutAction, {
-                loading: "Beenden...",
-                success: () => {
-                    router.push("/");
-                    return "Demo erfolgreich beendet.";
-                },
-                error: "Fehler beim Beenden der Demo.",
-            });
-        });
-    }
+export async function DemoBanner() {
+    const { user } = await getSession();
 
     return (
         <div className="fixed bottom-0 left-0 right-0 flex flex-col items-center justify-center bg-primary px-8 py-2 text-center text-primary-foreground">
-            <div className="flex flex-row items-center justify-center gap-2 text-sm">
-                <p>Derzeit ist der Demo Modus aktiv</p>
-                <Button
-                    className="text-sm text-primary-foreground underline hover:text-primary-foreground hover:no-underline"
-                    disabled={pending}
-                    onClick={() => {
-                        signOut();
-                    }}
-                    variant="link"
-                >
-                    Demo beenden
-                </Button>
-            </div>
+            {user && user.id === "demo" ? <DemoBannerExit /> : null}
             <div className="flex flex-row items-center justify-center gap-2">
-                <p>Haben Sie interesse? Dann füllen sie folgendes Formular aus</p>
+                <p>Haben Sie interesse? Dann füllen Sie folgendes Formular aus</p>
                 <Link
                     className="text-primary-foreground underline hover:no-underline"
                     href="https://forms.gle/rCLGkkNQoJQ51a7SA"
